extract control of music player to custom hook ... and remove favicon request
