+++ client/public/index.html
@@ -7,6 +7,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <meta name="theme-color" content="#000000" />
         <meta name="description" content="RAD music server" />
+        <link rel="icon" href="/">
         <!--<link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />-->
         <!--
         manifest.json provides metadata used when your web app is installed on a
