+++ client/public/manifest.json
@@ -1,23 +1,7 @@
 {
-    "short_name": "React App",
-    "name": "Create React App Sample",
-    "icons": [
-        {
-            "src": "favicon.ico",
-            "sizes": "64x64 32x32 24x24 16x16",
-            "type": "image/x-icon"
-        },
-        {
-            "src": "logo192.png",
-            "type": "image/png",
-            "sizes": "192x192"
-        },
-        {
-            "src": "logo512.png",
-            "type": "image/png",
-            "sizes": "512x512"
-        }
-    ],
+    "short_name": "RAD",
+    "name": "RAD Music Server",
+    "icons": [],
     "start_url": ".",
     "display": "standalone",
     "theme_color": "#000000",
