+++ client/src/components/MusicPlayer.js
@@ -1,109 +1,27 @@
-import React, { useRef, useState, useEffect } from "react";
+import React, { useEffect } from "react";
 
-import { calcDuration, constrain } from "../commons";
+import { calcDuration } from "../commons";
+import { useMusicPlayer } from "../hooks/MusicPlayer";
 
 const MusicPlayer = ({ name, track }) => {
-    const SPEEDS = ["1.0", "1.5", "2.0", "0.5"];
-    const audioElement = useRef(); // html5 audio element
-    const progressBar = useRef(); // div that shows the progress
-    const progressBarKnob = useRef(); // knob of the progress div
-    const animationRef = useRef(); // reference to animation frame used to update ui
-    const progressContainer = useRef(); // container of the entire slider
-    const speedRef = useRef(0); // current
 
-    const [isPlaying, setIsPlaying] = useState(false);
-    const [time, setTime] = useState({
-        current: 0,
-        max: 0
-    });
+    const {
+        isPlaying,
+        time,
+        audioElement,
+        progressBar,
+        progressBarKnob,
+        progressContainer,
+        togglePlayPause,
+        changePlaySpeed,
+        clickSeek,
+        seek,
+        changeTrack,
+        SKIP,
+        currentSpeed
+    } = useMusicPlayer();
 
-    const SKIP = 30;
-
-    const changePlaySpeed = () => {
-        speedRef.current = (speedRef.current + 1) % SPEEDS.length;
-    };
-
-    const refreshAnimation = () => {
-        cancelAnimationFrame(animationRef.current);
-        animationRef.current = requestAnimationFrame(whilePlaying);
-    };
-
-    const togglePlayPause = () => {
-        const prevValue = isPlaying;
-        setIsPlaying(!prevValue);
-        if (!prevValue) {
-            audioElement.current.play();
-            animationRef.current = requestAnimationFrame(whilePlaying);
-        } else {
-            audioElement.current.pause();
-            cancelAnimationFrame(animationRef.current);
-        }
-    };
-
-    // update the ui elements as the audio plays
-    const whilePlaying = () => {
-        setTime({
-            ...time,
-            current: audioElement.current.currentTime,
-            max: audioElement.current.duration
-        });
-        audioElement.current.playbackRate = SPEEDS[speedRef.current];
-        setProgress(audioElement.current.currentTime, audioElement.current.duration);
-        animationRef.current = requestAnimationFrame(whilePlaying);
-    };
-
-    // this is called when u click on either of the 2 seek buttons
-    const clickSeek = (forward) => {
-        const offset = forward ? SKIP : -SKIP;
-        const currentTime = audioElement.current.currentTime;
-        const changeTime = constrain(currentTime + offset, 0, time.max);
-        audioElement.current.currentTime = changeTime;
-        setTime({ ...time, current: changeTime });
-        setProgress(changeTime, time.max);
-    };
-
-    // this is called when u drag the slider knob or a point on the slider itself
-    const seek = (e) => {
-        const min = progressContainer.current.getBoundingClientRect().x;
-        const width = progressContainer.current.getBoundingClientRect().width;
-
-        // calculate percentage played based on the size of the slider
-        const percentage = (Math.floor(constrain(e.clientX - min, 0, width)) / width);
-        const changeDuration = percentage * time.max;
-        audioElement.current.currentTime = changeDuration;
-        setTime({ ...time, current: changeDuration });
-        setProgress(changeDuration, time.max);
-    };
-
-    // update ui elements to reflect the progress of the audio
-    const setProgress = (current, max) => {
-        const percentage = max !== 0 && !isNaN(max) ? (current / max) * 100 : 0;
-        progressBar.current.style.setProperty("width", percentage + "%");
-        progressBarKnob.current.style.setProperty("left", percentage + "%");
-    };
-
-    useEffect(() => {
-        const seconds = Math.floor(audioElement.current.duration);
-        setTime({
-            ...time,
-            max: seconds
-        });
-        setProgress(time.current, time.max);
-        audioElement.current.addEventListener("ended", () => setIsPlaying(false));
-        audioElement.current.addEventListener("error", () => {
-            setIsPlaying(false);
-            // auto get next audio to play
-            // ...
-        });
-
-    }, [audioElement?.current?.readyState]);
-
-    useEffect(() => {
-        togglePlayPause();
-        audioElement.current.play();
-        setIsPlaying(true);
-        refreshAnimation();
-    }, [track]);
+    useEffect(() => changeTrack(), [track]);
 
     return (
         <div className="sticky bottom-0 z-20">
@@ -218,7 +136,7 @@ const MusicPlayer = ({ name, track }) => {
                         type="button" onClick={changePlaySpeed}
                         className="rounded-lg text-xs leading-6 font-semibold px-2 ring-2
                         ring-inset ring-gray-200 text-gray-200">
-                        {SPEEDS[speedRef.current]}x
+                        {currentSpeed()}x
                     </button>
                 </div>
             </div>
